CIC part of petapm.
The CIC field on local PFFT domains are properly evaluated.

Warning: check why the total mass in region mesh doesn't match up
with the total mass of particles.

the total number of particles processed shall match with the total number
of particles. any no particles are outside of the regions they belong to.

The difference is too big to be explained by round off.