rework generation counting in SFR; preserve GAS ID.
Now every time we fork a particle, the parent's Generation is increased.
Also the particle ID of the child is set to a new value.

Be aware that care still needs to be taken such that a forked particle shall
never fork again.

We also infer the stellar generation from the gas particle mass rather than
from ID.

This means that the initial ID of gas particles are never changed.