fix the particle spawn code.
2 things;
Star_converted is no longer used in BH because we no longer remove gas
particles.

This is tricky though, because if after star formation the gas particle is less
massive than a BH seed, we will be left with a negative massed gas particle.
Shall warn about this at start up.

Also, we correctly set the ID of the new particle. It is important the new
particle id differ from the old because this is used to randomize the tree
nodes. The forked particle and the old particle has the same location!