Clean up the output list function a bit, by sorting the OutputListTimes array on read.
Rewrite the function for reading output times.

THIS BREAKS CONFIG: OutputListFilename -> OutputList

Yu Feng correctly observed that this function is crazy.
For one thing, it reads the file on every processor.
So instead we shall pass the output times as a comma-separated list, eg:
OutputList 0.1,0.2,0.3,0.5,1.0

These do not need to be ordered.

Thus we no longer need a separate output times file.

Also get rid of the arbitrary limit on the number of outputs, and get
rid of DumpFlag, which was not checked.

Further, sort the inputs on parsing, so we can clean up the
find_output_times function.