Speed up domain decomposition.
The domain decomposition was very slow relative to P-Gadget3.

It turns out that this was because we were recomputing the peano keys
multiple times in every domain decomposition (every time
domain_layoutfunc was called). Worse, this was done in some
hard-to-parallelize loops!

This patch fixes the problem by adding a Key member to struct
particle_data which is updated at the beginning of domain and used
throughout.

This speeds up the domain, but at a cost in memory. Future commits may
reduce the memory cost.