Add bad model for helium reionisation.
Add the ability to rescale the photoheating rates in a density dependent
way, following Becker 2011. This (sort of) models helium reionization
and allows us to produce different lyman alpha forest T_0 and gamma.

This is currently needed to get a lyman alpha forest that matches
observations from the simulation, but it is not a good model for helium
reionisation and it ultimately needs to be replaced.

Note that we convert from nHcgs (hydrogen number density cm^-3)
back to density (g cm^-3) in he_reion_factor, which includes a factor of
XH.

Forest paramters are compiled in even if SFR is off, as you may want to
change the helium heating even without star formation (although the most
common use case involves quick lyman alpha star formation being on).