Remove CPUSPEEDADJUSTMENT
This weighted the particle load per CPU by the ratio of:
gravcost/tree walltime.

The idea (I assume) is that slower CPUs can receive fewer particles.
However, there is only one binary, and there are so many syncpoints in
the code, that I can't see how this would ever work well in practice.