Add a pairwise short range gravity computer.
runtests.c does short range with tree and pairs. They seem to be
consistent.

Also added code from the bluetides-i grav tree code.
It gives the same result as here.

All these new code branches depend on the treewalk module. If that
one is fucked up we have no way of detecting it.