Do full domain decomp and tree rebuild before PM.
Previously we were drifting nodes and particles when marking
particles to regions, but this caused problems with very fast particles.
For fast particles the nodes were resized by Vel*dt_drift, a potentially
substantial amount causing excessive memory allocations.

A domain decmp and tree rebuild sycs all particles to the current time
and puts them on the correct processors. This is the easiest way to
deal with excessively large nodes. It also means the
tree is up to date and properly balanced, which avoids slow walks to
other processors (and may improve the PM time). Finally, it makes sure
the particles are properly wrapped to the box.

Also make sure the tree isn't rebuilt twice unneccessarily after a long
range force.