Reorder loop in force_tree_build_single so the logic flow and locking is clearer, and fix race conditions with a spinlock.
Because the parent node could get modified under us, the openmp force tree
build had a race condition. Fix this by using an array of SpinLocks
to protect modifications to Node.u.suns in the
tree_build main loop. The branches in the loop also get reordered to
make the locking clear.