Implement a small thread-local cache for the free Nodes.
I found that although parallelizing the treebuild sped up building, it
slowed down the tree walk. This is because the tree Nodes are generated
in the same order they are walked, so are very "cache-hot". In order to
avoid this, I implemented a cache for the free Nodes. threads will take
a free Node from the cache rather than the global list, thus ensuring
that neighbouring nodes are close together in the list.