Make local_topnode_data topNodes a local variable.
This makes it clear that the extra data members are only used in
determineTopTree. Thus the topNodes structure can be freed after the
results of this function are copied to the (global) topnode_data TopNodes
structure. The later walking of the domain topnodes can be done with
this smaller object.