Sync the integer timeline with the desired snapshot outputs.
Previously to this commit, snapshots were output by drifting particles
inside find_sync_point_and_drift from the end of the last timestep to
the desired snapshot output time.

However, this was inaccurate, because there should have been some
acceleration corresponding to the part of the timestep for which the
particles were drifted. Also, if the timestep in question was not a PM
timestep, there should have been some fraction of a PM timestep
accomplished. Once the snapshot is written and reloaded, PM_Ti_begstep
will be reset to the snapshot time and the acceleration from this part
of the timestep will be lost.

The fix for this is simple; ensure that the snapshots are output only at
the end of an integer timestep, after advance_and_find_timesteps.
To make sure that the snapshots still happen in a timely fashion, a
sync point is forced at the next integer output time.

We also enforce that this integer output marks the end of a PM timestep,
to make sure we get the GravPM accelerations. In rare circumstances this
could lead to very short PM timesteps, but this should not be a problem.

We maually ensure the kick does not advance past the PM timestep,
otherwise a PM step would never run again.

One minor unfortunate consquence is that a full timestep must now take
place between a TIMEOUT action and a snapshot being written. Note it is
still the case that a TIMEOUT action does not sync with the PM timestep.
We do not shorten the PM timestep when a TIMEOUT happens, because the
length of the timestep appears in the predicted SPH velocity, which
would then be incorrect.

THIS COMMIT CHANGES SNAPSHOT OUTPUT (slightly).