Do not do peano order in domain_maintain.
As well as saving time on the sort, this actually
makes the treewalk substantially faster!

The reason for this is unclear, since PH order
does improve cache efficiency.

It might be that it helps the prefetcher somehow?
Or it might be that domain exchange naturally puts
the active particles together, near the end of the array.