Cleaner time step bookkeeping.
Explicitly remember the current state of short range Kick and Drift.
of particles, and the current state of long range Kick and Drift.

This makes the logic in prediction cleaner (compensating difference
between drift and kick).

Move the first half kick before find time step, because it is finishing
the previous time step. The second half kick uses the new time step.