Build a list of particles of the desired type before calling petaio_build_buffer.
IO blocks are sorted by type at registration so that the buffer need
only be constructed once, rather than for every block. This saves some
IO time at the cost of memory.

Now enforce that in petaio_build_buffer selection != NULL.
This will fix the segfault encountered on icc (likely a compiler bug)
and lead to optimizations in later commits.