Totally refactor force_tree_build_single
Now we have a unit test we can do this more easily!

Split force_tree_build_single into multiple smaller functions, each with
several more comments. Make the locking clearer: almost always, each
node is now locked exactly once, and the nodes are locked and unlocked
in one place in the code. Make the handling of very close particles
much more robust: it is no longer random. Instead if a node is very
small, we place the new particle in the first free subnode.

Profiles show that the largest single cpu cost to this code is computing
the morton keys and the second is finding the TopLeaf.