Add stub for the better memory allocator.
The intention is to allow multiple allocators, and allocation
from both directions.

This would help cleaning up code sequences that are currently
limited due to we only have a single allocator and a single direction.

The code needs to be tested, but unfortunately cmocka doesn't work
on my box (many ref not found error during linking).