Remove forcetree Node MultipleParticles flag as it makes no sense.
1) It was always true (one only creates a tree node when it can contain
two particles). The only exception is if one somehow has an empty
TopNode, which the load balancing should prevent.
2) It was used to unconditionally open the node if it was false. But
this makes no sense, because if there are is only one particle in the
node, the node's mass and c of m will be the same as the particle, so
there is no need to open it!