Check for TIMEOUT and AUTO_CHECKPOINT events only on a PM step.
Before, if the time taken for a PM step was shorter than the last
step, it was possible for a TIMEOUT to occur on a non-PM step, changing
Ti_nextoutput, and then not occur on the PM step itself. The code then
signalled "stopping due to TimeLimitCPU", wrote a snapshot, changed
its mind and continued. While this was in some sense the correct thing
to do, it was a little confusing to the user. So now only TIMEOUT on a
PM step when we can do something about it. Similarly for
AUTO_CHECKPOINT, which makes little sense to signal except on a PM step.

Also re-order the events so that the STOP RIGHT NOW events take priority
over the "make a snapshot" events.