Move star-specific entries in particle_data into star_particle_data.
Currently contains FormationTime and Metallicity. Both are set on
creation and never altered. Metallicity is also in sph_particle_data
and FormationTime is also in bh_particle_data.

METALS is now always on (since it has small cost). It won't do much if
SFR is off. star_particle_data stores the same number of particles
as bh_particle_data