When forking a particle, only make it active if the parent is active.
It can happen that a BH is spawned from an inactive particle. If this
happens the Kick time will be out of sync and we crash. This doesn't
really matter, because BHs are moved to the potential center anyway,
but it is not ideal.

This fixes a crash at z=4.5 for the hydro example run for me.