Keeping TimeBinCount up to date was proving annoying, so Don't Do That.
We didn't update TimeBinCount properly after a domain exchange, or for
black hole formation, which occasionally led to eternal loops if a
TimeBin was empty. We don't really need it anyway: we just need the
minimum occupied timebin, and to print immediately after doing it.

Remove spurious fix from before.