Implement entropy limiting in predicted entropy.
That way the predicted entropy will not drop too fast, mirroring
the behaviour of the actual entropy in do_the_kick. This removes
a failure mode of the current code, where EntPred returns negative
and nans appear, hit by me under somewhat unusual circumstances.

A smaller fix would be to simply assert that predicted entropy is >= 0.