Re-do the displacements for gas even if they have the same power.
Shifting the gas particles by a full grid cell without changing their
velocities was leading to a small but noticable loss in gas power.

Before it was tolerable as both the gas and DM were only off by half a
grid cell.

Also fix a bug where the IDs were re-used for gas and DM.