Remove the twisted logic about keeping SofteningType.
It appears that the entire logic there was to save freaking 4 bits;
making the code very hard to parse.

Now we only keep MaxSoftening of each node and if the value represents
all softening -- the node is opened when it is not --
which is almost always the case when there is adaptive smoothing.

Added function get_softening shall go somewhere else or get a better name.