Copy qsort from glibc so we can use pre-allocated tmp memory.
glibc's qsort is actually an out-of-place merge sort, which means it
does a malloc() using system malloc. This plays havoc with our
carefully-designed memory subsystem, and causes fragmentation.
The easy solution is to inline glibc's merge sort and use the
pre-allocated memory buffer, which we need anyway for the parallel part
of the algorithm.