Use indirect sorting also during the merge phase
If the struct to be sorted is large, glibc uses an indirect sort.
Make this indirect sort be applied during the merge phase as well, so we
only do the memcpy at the very end. Final memcpy is serial.

Note: for some array sizes the time is dominated by the final serial
memcpy. In this case the algorithm performs the same as the serial algorithm,
rather than worse.

With this commit, there should be no situation where it is better to use
libc qsort.