Split garbage into two types: sorting and compacting.
Compacting is as before but with a new sweep algorithm which memmoves
the particles over the garbage. This function is O(n) instead
of O(n log n), because it does not sort. It also preserves the
ordering of the particles, instead of tail copying, which may
improve cache usage.

Sorting is the old peano sort in domain, but also moving garbage to the end,
followed by a sort on the slots. It is used only in
domain_decompose_full.

This somewhat reduces memory usage, but note there was a bug before:
exchange did not know about the memory allocated by qsort.

I added a per-type garbage counter to count the number of
particles that need to be removed. If this counter is small, we skip
compactifying the slot array, as before.

Also use the GC event to signal rebuilding the tree on snapshot write if there was garbage.