Initialise SlotsManager ptrs
This ensures that we do not pass a null pointer to memmove on the first
pass, which is UNDEFINED BEHAVIOUR! This memmove was of zero size, so
should have done nothing, BUT compilers can do weird things when aggressively
vectorising and encountering null pointers. This probably caused my
optimization-dependent crash with icc.