Always check for drift/jump of BH particles.
Previously we may not immediately drift a BH particle even if
Jump flag is set if the Ti is already current(due to drifting induced
by pair interaction with other particles). As a result
BHs are still not fully synced to the pot min.

If the logic here is correct it should have cured the problem.

Again, this brings up the question -- how important is it to ensure
that all changes to Position is centralized to the same routine?

And why is it called a drift?