Optimize treewalk by removing useless copy.
The treewalk code contained a useless copy from WorkSet to PrimaryTasks:
PrimaryTasks appears to be a partial implementation of a feature which
would allow the treewalk to restart from a node other than -1 (which is
a useful feature). However, in the current code it does nothing.

Remove it, but leave the code in and commented for future reference.