Make sure PM step is not longer than the longest particle timestep, fixing a rare edge case in the timestep code.
If the PM step was longer than the longest particle timestep, and
some of the particles in the longest timestep dropped to a longer
timestep, they could fail to be active on the next PM step, causing the
integer timeline to get out of sync. Fix this by enforcing that the PM
step is never longer than the longest non-empty timestep.