Remove MPI_Barrier from message(0, ...)
I think the main purpose was so that the time messages in domain were accurate, but
and we have better ways of knowing how long domain took nowadays.

The MPI_barrier also makes multiline messages unreasonably expensive and
can lead to irritating hangs if one writes a debug print which is conditional
somehow.