Script to convert bigfile to Gadget format 3 (#232)
This adds a python script to convert a bigfile snapshot to a gadget format 3 HDF5 snapshot. The use-case is people with existing analysis pipelines, so they can try out MP-Gadget without needing to switch snapshot format immediately.

At the moment it is pretty bare-bones: it should do Gadget format 2 as well, and support the reverse: conversion of Gadget formats 2,3 to bigfile.

Travis tests that it works and completes conversion: the results look roughly correct. I have also sent this to someone with an existing pipeline and they have not complained.