Move make_particle_star out of cooling loop, remove thread contention.
This allows us to locklessly convert a particle into a star.
In particular it means no contention on the atomic with
QuickLymanAlphaProbability > 0, which was causing me problems.

We do this by making a threaded list of the new stars, and by counting them.

There is still splitting of particles in the loop, with an atomic on
incrementing NumPart, but because fewer stars form with QuickLymanAlpha
off, this seems ok. It was also tricky to remove it and keep the wind
model working ok.