Various moves and API changes to fit into the rest of the code
This initialises the cooling module directly from params.c and removes
its entries from allvars. It also moves the metal cooling to the same
file as the UV fluctuations.

We have new parameter file options for the SelfShielding and the cooling
and recombination rates, still set to their old defaults