Add a glass solver to Gen-IC.
If MakeGlass is set, the glass solver is invoked to produce a glass particle
distribution in place of the original grid distribution.

The current scheme uses a damped inverted gravity force.

The numerical scheme was tested with fastpm-python and it gives
a k**4 power spectrum from k=0 up to the nyquist when B=2.

glass.c will write the power spectrum per iteration to the output directory.
The file name contains -glass, and the random seed of the glass.

The current random field is not MPI size invariant.

Someone smarter shall fix it.