Initialize Entropy on reload with DENSITY_INDEPENDENT_SPH is off
This uses the InternalEnergy saved array to initialize the Entropy
variable. Before Entropy was being saved and load with
DENSITY_INDEPENDENT_SPH on but not with it off.

The easiest thing to do is to save and load Entropy unconditionally, but
that would break compatibility with existing snapshots. So to minimalize
the behaviour change with #ifdefs now use InternalEnergy and do a
conversion in both cases.