Instead of creating Injected_BH_Energy and adding it in the SFR code, just add the BH feedback energy directly into blackhole.c
This will change output - the previous code was altering the cooling
time of stars on the effective equation of state so that if they were
hit with black hole feedback they always cooled to the effective energy
immediately. This doesn't really make sense and also doesn't make a lot
of difference (particles on the equation of state cool to the effective
energy pretty quickly anyway).

So just remove it: black hole feedback energy is now added directly to
the gas particle. This decreases the star formation rate by perhaps
0.5%.