Clean up initializing EgyWtDensity.
Instead of checking for it being < 0 in density, and doing different
things depending on whether it is the first loop or not, set it
to Density in init.c before the first loop.

Also remove the confusing split functions for density. Use arguments.
Add an argument that means "compute the energy weighted density"
and use it appropriately in init.c so that it is clearer how
EgyWtDensity is initialised.