Use the same velocities in snapshots as in Gadget-2/3 and as in ICs.
This introduces a new header flag specifying whether Gadget-formatted
velocities are used, so that downstream analysis codes can tell the
difference. By default the velocities output are the same as those read
in the ICs. If the flag is not present in the ICs then it is assumed
the ICs have the 1/sqrt(a) factor, but the snapshots do not.