Refactor sfr code to avoid scratch data during neutral fraction computation.
The neutral fraction for particle data output called the star formation
rate function, which may use scratch data. This does not work after
FOF because the FOF code exchanges particles. However, the neutral
fraction does not actually use the star formation rate. Refactor the
code so it does not try to compute it, which also makes the SFR code
simpler.