Set black hole timestep to timebinlimit, not < timebinlimit
The timestep of the black hole was being set to at most the smallest
timestep of any neighbouring SPH particle, but it does not make sense to
subsample the black hole timestep, so set it to exactly the smallest
timestep of any neighbouring SPH particle. Rename the variable so it
is clearer what it does.