Remove the sound speed check for black hole mergers
Because black holes are not dissipative, it is easy for them to
fail this speed check. Before we did not notice this, because black hole
drifting was broken anyway, so there were no mergers.

Now what happens is that multiple black holes, repositioned to the
minimum potential of the halo, all have exactly the same position, but
different velocities, and cause the tree evolution to break down!