Never prune empty top-level nodes from the force tree.
It is very rare that top-level nodes are empty, but it can happen if the
code is over decomposed. If it did, the assumption that a top-level node
has 8 daughters was broken. Because the removed node had not been
properly zeroed, this could lead to reading an int as a double and thus
NaN appearing in the force tree.