Remove the black hole velocity check on mergers
This was a bad idea, because the black holes are repositioned, but their
velocities are not zeroed, and they are not dissipative.
Over time it will create a clump of black holes at identical positions
in the center of the halo, unable to merge. This will very
much slow down the code.