Fix force_update_hmax
I realised that force_update_hmax contains a serious flaw: the hmax of a
top-level node is set to the hmax of the first particle to reach it, not the
actual maximum particle. Then this hmax is (wrongly) exchanged.
It rarely matters for top-level particles, but it is a bug.