Keep swallowed black holes around after swallowing.
This is to allow us to keep track of mergers directly.
The particles (and their SwallowID) are saved to the snapshots.
They do not move, even to drift (so they do not participate in exchange),
and they are not added to the tree or the active list.

Since on PM steps we do not use the active list, we still need to explicitly
exclude them from the PM grid and from density calculations (so that
their saved Hsml values reflect the meaningful Hsml from before their last
merger). The short-range tree force on them is still computed, but there
are always few BHs, so it doesn't matter much.

Adds two new fields to the snapshots for BHs: Swallowed and SwallowID.

Fixes: #67