Do not ignore wind particles when generating wind energy.
When we are generating new wind particles, we were ignoring existing
wind particles. This means that the wind is generated by whichever
new star happens to meet it first.

With OJFT10 winds the wind velocity depends on the local velocity
dispersion, so this meant that the wind velocity had a random
(timing-dependent) component. This is I think why the hydro example is
so sensitive to the number of threads.

To fix this, allow all existing wind particles to receive new wind
energy, and simply recharge their DelayTime. Not only is this
numerically more sensible, it seems fairly physical too.