Modify the way black hole heating is added so it doesn't affect DtEntropy
This avoids DtEntropy becoming unreasonably large around the black hole
and shrinking the timesteps. The same heating is included, but it is
impulsive, so does not affect the predicted entropy used in density and
hydro. This impulsive heating meant that the entropy prediction (which
is supposed to include adiabatic heating) was not accurate.