Do not set Swallowed for gas particles
We were setting the Swallowed attribute for gas particles once they were
near enough a BH to be accreted. They would then be set to zero mass
later, when we were sure which BH to accrete them onto. Rarely it would
happen that the gas particle was accreted and then the BH merged in the
same timestep, and is thus not active for the second accretion treewalk,
leaving a strangely inconsistent swallowed but non-zero mass particle
around.

Change this so that only SwallowID is set, which is a temporary. If the
BH goes away, it is discarded and the particle persists as before.

Also for ease of algorithm, set the temporary to store BHID + 1, not
BHID.