Change the meaning of hmax.
After this commit, it is the amount of extra node length we
should allow for symmetric tree walks. This is what it is actually used
for and should make tree walking a bit faster. More importantly, force_update_hmax
is much faster, because most particles are fully inside the node and
thus do not change hmax.