Do cooling after kicks using a Strang splitting approach.
In Gadget-2 the cooling works out DtEntropy at acceleration time
and then adds DtEntropy to Entropy at kick time. However, especially in
the presence of BH heating this may lead to unphysically short
timesteps as the predicted entropy is severely mis predicted.

For this reason Gadget-3 has switched to directly changing the
entropy after the kick. This behaviour corresponds to a more general
Strang splitting approach, as explained by Volker.