Revert ce3e0e2a84 due to major performance loss for threaded code.
ce3e0e2a84:
    Replace final alloca calls in treewalk

    Use instead allocations on the ta malloc heap. This removes a
    substantial source of thread-local stack allocations.

Removed thread-local stack allocations. This had no effect on a pure
MPI run but costs us a factor of 2 (!) in threaded performance, making
the threaded code half as fast as the MPI code.

I am not sure why this is. I suspect it may be that the stack variable
can be optimized out entirely by the compiler, but the heap variable
cannot. It may also be some effect of the processor caches.