Factor out kick and drift times into structure.
The kick times are the same for all particles in a timebin. We can
factor that out and thus only compute the kick factors once each
halfstep (and once more for predicted velocities). This avoids the
interpolation, which was expensive as it needed an exp() and was
O(particles) rather than O(active particles).

Note we advance kick times for small bins,
even if there are no particles in them. This keeps the kick
times synchronized.

Leave the old computation in for debugging.