Revert "split to multiple jobs"
This reverts commit f870a481163c4a894e5f6fa063e64728daffe1ab.

Passing binaries as artifacts in the current form is probably not easy,
so we'll continue to use a single job with sequential steps.

(We probably only need MP-Gadget and MP-NGenIC ..) Unlikely worth
the effort to figure it out. Unless we drastically increase the
number of E2E test scenarios (like example/travis), which I don't
think will happen any time soon (@sbird)