Implement improved predicted density from Gadget-4.
This replaces the exp() from Gadget-2, which was pathological at long
timesteps and so in practice unused, with a linear forward prediction
also using DivVel. This is only used in the hydro code: other parts of
the code using the unpredicted density for inactive particles are:

- Max BH heating temperature for entropy -> temperature conversion.
- metal_return for weighting the SPH mass return.

Neither of these are really important.

Density is also used for the entropy->temperature conversion factor
in cooling_qso_lightup.c, but this is only run when all particles are
active.