Fix the sibling fixup in tree build which was sometimes accessing data outside the local tree.
Instead of blindly using the Node one before the current one in the
tree, actually walk the parent's child array and find the true child.
I think this was generally harmless because the Node one before was
unlikely to have this one as a sibling, but it was a data race.