Use a DM-only tree for doing FOF linking and gas for He-reion.
The DM-only tree makes FOF linking much faster when there are many
stars. The gas only tree for He reion is unnecessary, but
it makes the memory management easier and would only be built on a PM
step.

A pleasant side-effect is that this removes a bunch of
memory management special cases.