Implement the push-down behaviour from Gadget-4
If all particles in a bin go to the bin below, we do not need to re-run
the accelerations (because they are the same). Because a hierarchical
acceleration step needs 3x the acceleration evaluations of a regular
step, it is advantageous to push all particles down a timestep if more
than 1/3 of them need to be pushed down.