Fix an edge case when all active gas particles become stars
In this case the code to find the minumum hydro timestep has nothing
to do, and so we end up with the minimal timestep being invalid, and
so the advance code doesn't work.

Fix this by reading from the hydro timestep of the active star
particles. We may still have useless timesteps if the next timestep also
has no gas in it, but that's a rare enough problem.