Fix to metal species return
The previous model confused stellar yields (fractional change to the
metal abundances in the star) with mass loss (total amount of metal lost
to the ISM). For metal species these two are about the same, but for H
and He they are not (because the yield is about how H is converted to a
metal, the H table can be negative). This was messing up the Metals
output table, but I don't think had any dynamical impact.

Thanks to Nishi Khandai for finding the issue with Metals.