Fix a case where mass conservation could be broken when a small BH accretes a gas.
The edge case is:
- A seed mass black hole accretes gas (this is fairly rare!)
- The gas mass, gM is added to O->acMtrack
- BHP(Mtrack) = BHP(Mtrack) + O->acMtrack
(note the dynamical mass, P.Mass for this BH is SeedBHDynMass here)
if(BHP(Mtrack) > SeedBHDynMass)
        BHP(Mtrack) = SeedBHDynMass
But here mass is lost! Mtrack is supposed to track the real physical
mass in the box. We started with BHP.Mtrack + gM and we finished with
SeedBHDynMass, potentially losing (BHP.Mtrack + gM - SeedBHDynMass).

This would be ok if the gas mass is constant, and SeedBHDynMass is a multiple of gM,
as Mtrack == gM initially. However, mass return from star formation
means this will not generally be true.