Fix entropy limiter in the prediction code.
There is code in the predicted entropy to prevent the predicted entropy
being too low. This can happen when the gas is cooling very quickly in
the cooling code, but the DtEntropy, which contains the entropy change
from volume effects, is not changed. In extreme edge cases the gas could
have negative entropy.

This limiter was previously applied only for gas whose timestep was
predicted forward (ie, gas before the current time). This corresponds to
gas which is cooling from adiabatic effects. Half the time the
gas will be predicted backwards. If the gas is being heated (compressed)
by adiabatic effects (such as from a BH kinetic feedback event) we might
again need the limiter.