Fix merger mass calculation if a seed black hole accretes multiple non-seed-mass black holes in a single timestep.
BH A is < SeedBHDynMass and BH B is > SeedBHDynMass.
BH A accretes BH B.

We want the new combined masses to be:
A.Mass = A.Mtrack + B.Mass
A.Mtrack = SeedBHDynMass

Before this commit we did:
temp.Mtrack = B.Mtrack (note that this is > SeedBHDynMass)
temp.Mass = B.mass + A.Mtrack - SeedBHDynMass

A.Mass += temp.Mass = B.mass + A.Mtrack
(as A.mass was previously SeedBHDynMass)
A.Mtrack = temp.Mtrack

which is correct.

However, if BH A also accretes BH C *in the same timestep* we will get

temp.Mtrack = B.Mtrack + C.Mtrack (this is fine)
temp.Mass = B.mass + C.mass + 2 * (A.Mtrack - SeedBHDynMass)

and then:
A.Mass += temp.Mass
A.Mass = B.Mass + C.Mass + 2 * A.Mtrack - SeedBHDynMass

which means we have lost A.Mtrack - SeedBHDynMass!

After this commit we do:
temp.Mtrack = B.Mass

or, if BH C is accreted:

temp.Mtrack = B.Mass + C.Mass

and then we add it to the main variables:
A.Mtrack += temp.Mtrack
A.Mass = A.Mtrack
A.Mtrack = SeedBHDynMass

which is correct