Simplify handling of BH mass accretion in the seed mass regime.
I can now simplify the mass of if statements, we need to accrete mass
onto BH A from BH B.

If B.Mass < SeedBHDynMass we need to accrete B.Mtrack

If B.Mass > SeedBHDynMass we need to accrete B.Mass

If A.Mass == SeedBHDynMass we accrete onto A.Mtrack

If A.Mass > SeedBHDynMass we accrete onto A.Mass

Previously this was 4 if statements in ngbiter. Now we just make
a temp.Mass with all the masses we want to accrete in ngbiter (ie, the
first two cases) and reduce and then decide in postprocess which variable to accrete onto.

This does not fix any bugs that I can see, but it is easier to read