Use a temporary for SwallowID
So that non-swallowed BHs do not have SwallowID set next timestep.

Consider this example:
BHs A,B,C, with A.ID < B.ID < C.ID.
A is swallowed by B, but B is swallowed by C in a single timestep.
In this case B will be swallowed and A will not be swallowed.

However, we would also set A.SwallowID = B.ID in the first treewalk.
Because SwallowID was persistent, BH A would be unable to merge or
swallow another BH until it found a BH with a higher ID than B.
Usually this is fine, because C is nearby and has a higher ID than B!
But it might be that A moves away from C and is unable to interact with
another black hole until it finds D with D.ID > B.ID.

I knew about this case and even wrote a comment about it!
This commit fixes the edge case by using a temporary SwallowID and only setting A.SwallowID
once the swallow actually happens.