Change the inttime_t type
Previously the integer time was just a 32-bit integer and we used bit
tricks to divide this into a snapshot piece and a timebin. But we began
to need more timebins for long runs, and the code was a bit confusing.

This makes inttime_t a struct with two 32-bit variables. The timebin
hierarchy is thus extended to 30 bins, instead of 20. Because it is no
longer a primitive type, and this is not C++, we need special functions
for comparison and addition. Most of the patch is the downstream changes
to make this work.