Attempt to fix auto checkpointing feature on Frontera
The latest intel compiler on Frontera is mysteriously setting
HCIManager->timer_begin to some large number. We hypothesise that
when hci_now sets a local function variable to the current time and then
broadcasts it, the compiler is freeing the function stack and over-writing it,
broadcasting the wrong variable. This only happens on Frontera.

Attempt to fix this by storing the MPI_Wtime in a long lived variable of
HCIManager before broadcast. Also reoder HCIManager struct so that there are no
gaps and the compiler is less tempted to be clever.