Unconditionally use a separate temp to store accelerations in gravtree
We read FullTreeGravAccel in grav_short_copy and modify it in
grav_short_reduce. This was dangerous and when the buffer filled up
could result in a wrong partial FullTreeGravAccel being used for some
particles.

Also use full_particle_tree_flag to decide when to update
FullTreeGravAccel, which incidentally fixes a bug where the potential
was not updated on all PM steps if the particle table contained garbage.