Fix possible deadlock
When we do slots_reserve in sfr we are not doing it collectively.
So if we need to rebuild the tree in wind, we may do it on only 1
process and hang. Do not free the tree in slots_reserve, copy memory
instead.

Revert "Since we might need to rebuild the tree in wind anyway, just free it on needing to reserve extra slot memory, which is rare"

This reverts commit 53f4ef3e1da5f863e0ef5a6c9aaf03189b33f459.